requirements.txt
discord.py>=2.0.0
python-dotenv>=0.19.0